Moved the functionality of bookmarking from Resource detail card to Resources Page
